requirements.txt
requests==2.25.1
toml==0.10.2
Twisted==21.2.0
ujson==4.0.2
python-dateutil==2.8.1
durations==0.3.3
